pyproject.toml
[project]
name = "aibi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno[lancedb,openai,sqlalchemy]>=2.5.15",
    "fastapi>=0.135.3",
    "groq>=1.1.2",
    "lancedb>=0.30.2",
    "mysql-connector-python>=9.6.0",
    "openai>=2.31.0",
    "openrouter>=0.8.1",
    "pandas>=3.0.2",
    "pgvector>=0.4.2",
    "plotly>=6.6.0",
    "psycopg2-binary>=2.9.11",
    "pymysql>=1.1.2",
    "pyrefly>=0.60.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0",
    "tantivy>=0.25.1",
    "uvicorn>=0.44.0",
]
